Replace TFeldmanCousins with table from FC paper
ROOT.TFeldmanCousins.CalculateUpperLimit(nb,nb) returns the upper limit on ns
given exactly nb observed events on an average background of nb, not the
average upper limit across an ensemble of experiments with ns=0 and nb exactly
known. The answer is the same as Feldman & Cousins Table XII for nb=0, but not
for any other case; in particular it's off by a factor of 2 for nb=0.5. To
avoid mistakes (and a ROOT dependency), use an interpolation of the table
instead, raising and error for nb>15.

git-svn-id: http://code.icecube.wisc.edu/svn/sandbox/jvansanten/projects/2015/gen2_analysis@160261 16731396-06f5-0310-8873-f7f720988828