Use saturated log-likelihood ratio rather than log-likelihood
Since the raw log-likelihood scales with the number of entries in each bin,
likelihood contributions can have vastly different magnitudes. This leads to
significant round-off error when the sum is taken. The ratio to the peak
likelihood, on the other hand, is more uniform, and so the sum is more stable
this way.

git-svn-id: http://code.icecube.wisc.edu/svn/sandbox/jvansanten/projects/2015/gen2_analysis@160599 16731396-06f5-0310-8873-f7f720988828